Adds invoice download functionality

+++ database/migrations/2015_11_05_175811_create_payments_table.php
@@ -14,7 +14,7 @@ class CreatePaymentsTable extends Migration
     {
           Schema::create('payments', function (Blueprint $table) {
             $table->increments('id');
-            $table->float('amount');
+            $table->decimal('amount',10,2);
             $table->string('bank');
             $table->string('confirmation_code');
             $table->date('date');
